メッセージへの返信に関するコードを移動
