+++ internal/messages.js
@@ -18,6 +18,8 @@
  */
 
 const { Client, Message } = require("discord.js");
+const axios = require('axios').default;
+const { strFormat, LANG } = require("../util/languages");
 
 /**
  * 自動応答のパターン。
@@ -159,14 +161,125 @@ class ClientMessageHandler {
     /**
      * メッセージを受け取ったときの処理。
      * @param {Message} message メッセージ
-     * @returns {Promise<boolean>} メッセージに反応したかどうか
+     * @returns {Promise<void>} メッセージに反応したかどうか
      */
     async handleMessage(message) {
+        if (message.author.bot) {
+            return;
+        }
+
         const guild = message.guild;
         if (guild == null) {
-            return false;
+            return;
         }
-        return this.getGuildMessageHandler(guild.id).handleMessage(message);
+
+        const done = await this.getGuildMessageHandler(guild.id).handleMessage(message);
+        if (done) {
+            return;
+        }
+
+        await replyBetterEmbedUrl(message);
+    }
+}
+
+//!function
+async function getRedirectUrl(shortUrl) {
+    try {
+        const response = await axios.head(shortUrl, {
+			maxRedirects: 0,
+			 validateStatus: (status) => status == 301 || status == 302
+		});
+        const redirectUrl = response.headers.location;
+        console.log(LANG.discordbot.getRedirectUrl.redirectURL, redirectUrl);
+        return redirectUrl;
+    } catch (error) {
+        console.error(LANG.discordbot.getRedirectUrl.error, error.message);
+		return `${LANG.discordbot.getRedirectUrl.error} ${error.message}`
+    }
+}
+
+/**
+ * TODO: 簡略化
+ * vxtwitter.com, fxtwitter.com, vxtiktok.com の URL を返信する可能性がある。
+ * @param {Message} message メッセージ
+ * @returns {Promise<void>} メッセージに反応したかどうか
+ */
+async function replyBetterEmbedUrl(message) {
+    const urls = message.content.match(/https?:\/\/[^\s]+/g);
+
+    if (urls) {
+        for (let url of urls) {
+            if (url.includes('twitter.com') || url.includes('x.com')) {
+                if (url.includes('vxtwitter.com') || url.includes('fxtwitter.com')) { //ignore vxtwitter.com and fxtwitter.com
+                    return;
+                }
+                await message.react('🔗'); // リアクションを追加
+
+                const filter = (reaction, user) => user.id == message.author.id && reaction.emoji.name === '🔗';
+                const collector = message.createReactionCollector({ filter, time: 30000 });
+
+                collector.on('collect', async (reaction, user) => {
+                    const modifiedURL = url.replace('twitter.com', 'vxtwitter.com').replace('x.com', 'vxtwitter.com');
+                    let fxmsg = strFormat(LANG.discordbot.messageCreate.requestedBy, [user.username]) + `\n${modifiedURL}`;
+                    message.channel.send(fxmsg)
+                        .then(sentmsg => {
+                            message.reactions.removeAll().catch(e => {
+                                console.error(strFormat(LANG.discordbot.messageCreate.reactionRemoveErrorConsole, [e.code]));
+                                let errmsg = '\n' + strFormat(LANG.discordbot.messageCreate.reactionRemoveError, [e.code]);
+                                sentmsg.edit(`${fxmsg}${errmsg}`);
+                            })
+                        })
+
+                    collector.stop();
+                });
+
+                collector.on('end', (collected, reason) => {
+                    if (reason === 'time') {
+                        // TIMEOUT
+                        message.reactions.removeAll();
+                    }
+                });
+            }
+            if (url.includes('vt.tiktok.com') || url.includes('www.tiktok.com')) {
+                if (url.includes('vxtiktok.com')) {
+                    return;
+                }
+                await message.react('🔗'); // リアクションを追加
+
+                const filter = (reaction, user) => user.id == message.author.id && reaction.emoji.name === '🔗';
+                const collector = message.createReactionCollector({ filter, time: 30000 });
+
+                collector.on('collect', async (reaction, user) => {
+                    console.log(strFormat(LANG.discordbot.messageCreate.beforeUrl, [url]));
+                    if (url.includes('vt.tiktok.com')) {
+                        url = await getRedirectUrl(url);
+                    }
+                    console.log(strFormat(LANG.discordbot.messageCreate.afterUrl, [url]));
+                    if (url.includes('Error')) {
+                        message.channel.send(LANG.discordbot.messageCreate.processError + "\n" + "```" + url + "\n```")
+                    }
+                    const modifiedURL = url.replace('www.tiktok.com', 'vxtiktok.com');
+                    let fxmsg = strFormat(LANG.discordbot.messageCreate.requestedBy, [user.username]) + `\n${modifiedURL}`;
+                    message.channel.send(fxmsg)
+                        .then(sentmsg => {
+                            message.reactions.removeAll().catch(e => {
+                                console.error(strFormat(LANG.discordbot.messageCreate.reactionRemoveErrorConsole, [e.code]));
+                                let errmsg = '\n' + strFormat(LANG.discordbot.messageCreate.reactionRemoveError, [e.code]);
+                                sentmsg.edit(`${fxmsg}${errmsg}`);
+                            })
+                        })
+
+                    collector.stop();
+                });
+
+                collector.on('end', (collected, reason) => {
+                    if (reason === 'time') {
+                        // TIMEOUT
+                        message.reactions.removeAll();
+                    }
+                });
+            }
+        }
     }
 }
